refactor(LottoMachine): LottoMachine WinningLotto와 분리
